feat: include column numbers in `git grep` results
This would be a breaking change because `git grep` only acquired
the `--column` switch in 2.16.0 (released 2018-01-17). That's about
6.5 years ago now, so hopefully this wouldn't cause too much pain.
Nevertheless, I've implemented a runtime check to see whether `--config`
is supported.

At the same time, fixed a bug in the `errorformat` string, which
included a space before the message which would prevent lines from
matching and being correctly picked up by the quickfix listing.