fix: make `diff` mode work with filenames with spaces in them
When Git emits diff headers for filenames with spaces in them, it uses
this crafty code:

    case DIFF_SYMBOL_FILEPAIR_PLUS:
            meta = diff_get_color_opt(o, DIFF_METAINFO);
            reset = diff_get_color_opt(o, DIFF_RESET);
            fprintf(o->file, "%s%s+++ %s%s%s\n", diff_line_prefix(o), meta,
                    line, reset,
                    strchr(line, ' ') ? "\t" : "");
            break;
    case DIFF_SYMBOL_FILEPAIR_MINUS:
            meta = diff_get_color_opt(o, DIFF_METAINFO);
            reset = diff_get_color_opt(o, DIFF_RESET);
            fprintf(o->file, "%s%s--- %s%s%s\n", diff_line_prefix(o), meta,
                    line, reset,
                    strchr(line, ' ') ? "\t" : "");
            break;

- https://github.com/git/git/blob/2c2ba49d55ff26c1082b8137b1ec5eeccb4337d1/diff.c#L1515-L1528

The "a/" and "b/" cases are the same, but in our case our regexp is
looking for the "b/" case, so let's focus on that. Say we have a diff to
a file like "foo/bar/baz.txt", Git will print something like:

    +++ b/foo/bar/baz.txt

But if the file has spaces in the name -- for example, "baz with
spaces.txt" -- Git will print a tab at the end:

    +++ b/foo/bar/baz with spaces.txt\t

As this is normally invisible, you don't notice, but vcs-jump falls
afoul of it because we capture the whole string after "b/" as the
filename (ie. "foo/bar/baz with spaces.txt\t"). This in turn leads it to
make a broken quickfix file:

    foo/bar/baz with spaces.txt\t:100:some line

Because there is no such file, jumping to it in Neovim will just land
you in an empty buffer.

I imagine that the reason for the Git behavior is to facilitate some
machine-based parsing, but in practice I would argue that the handling
is at best a bit fiddly to work with and at worst broken:

- "foo bar" will be emitted as "foo bar\t"
- "foo\t" will be emitted at "foo\t"
- "foo bar\t" will be emitted as "foo bar\t\t"

in other words, consumers can't just mindlessly strip off the trailing
tab; doing so would break the "foo\t" case. Instead, they have to check
for a trailing tab _and_ also for space, and only strip the tab if
a space is found.

In any case, I am not going to overthink this. I could add logic to
strip a trailing tab if and only if I also see a space, but I am just
going to keep things simple and strip the trailing tab as soon as I see
it. This will make the common case (spaces with filenames in them) work
with minimal fuss; if anybody ever comes complaining to me that the
filename-with-no-spaces-but-trailing-tab-is-actually-on-the-filesystem
case is now broken, then I'll implement the more complicated handling.